pyproject.toml
[tool.poetry]
name = "gitlab-dashboard"
version = "0.1.0"
description = "Create dashboard for checking develop activity of develop team using gitlab."
authors = ["sUeharaE4"]

[tool.taskipy.tasks]
start = { cmd = "cd src && streamlit run main.py", help = "launch server" }
test = { cmd = "pytest -vv --durations=0 --log-cli-level=10", help = "runs all tests" }
test_cov = { cmd = "coverage run -m pytest -vv --durations=0 --junitxml=build/tests/result.xml --cov=src --cov-report=xml:build/tests/coverage.xml --cov-report=html:build/tests/htmlcov --html=build/tests/report.html", help = "runs all tests and make coverage report" }
sphinx = { cmd = "task sphinx_apidoc && task sphinx_build", help = "make sphinx doc" }
check_style = { cmd = "task check_flake8 && task check_flake8_with_output && task check_pyright && task check_mypy", help = "check code style" }
check_complexity = { cmd = "cd src && docker run --rm --env CODECLIMATE_CODE=`pwd` --env CODECLIMATE_DEBUG=1 -v `pwd`:/code -v /var/run/docker.sock:/var/run/docker.sock -v /tmp/cc:/tmp/cc codeclimate/codeclimate analyze", help = "check_code complexity use docker image of code climate. this command needs docker." }
sphinx_apidoc = { cmd = "sphinx-apidoc -M -f -o ./build/docs/source ./src", help = "sub commands for make sphinx doc" }
sphinx_build = { cmd = "sphinx-build -b html ./build/docs/source/ ./build/docs/build/", help = "sub commands for make sphinx doc" }
check_flake8 = { cmd = "flake8 ./src --docstring-convention numpy", help= "sub commands for check_style" }
check_flake8_with_output = { cmd = "flake8 ./src --docstring-convention numpy --format=pylint --output=./build/check_style/flake8.txt", help= "sub commands for check_style" }
check_pyright = { cmd = "pyright ", help= "sub commands for check_style" }
check_mypy = { cmd = "mypy ./src/**/*.py", help= "sub commands for check_style" }

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3

[tool.pyright]
pythonVersion = "3.9"
include = ["src"]
exclude = ["tests"]

[tool.mypy]
python_version = "3.9"
exclude = ["tests"]
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.poetry.dependencies]
python = "^3.9"
taskipy = "^1.8.1"
streamlit = "^1.11.1"
python-dotenv = "^0.19.2"
streamlit-aggrid = "^0.2.2"
python-gitlab = "^2.10.1"
altair = "^4.2.0"
stqdm = "^0.0.4"
notebook = "^6.4.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.910"
black = "^21.7b0"
flake8 = "^3.9.2"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"
pytest-randomly = "^3.10.1"
pytest-html = "^3.1.1"
Sphinx = "^4.1.2"
sphinx-autodoc-typehints = "^1.12.0"
sphinx-rtd-theme = "^0.5.2"
flake8-docstrings = "^1.6.0"
isort = "^5.10.1"
pyright = "^0.0.13"
mistune = "^2.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
